+++ mangaScraper/Docs/ChangeLog.md
@@ -1,3 +1,4 @@
 ## V 0.1.4
  - Change Log Added
  - Added customizable defaults 
+ - Changed Langauge to Language